% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoSoilQuery-package.R
\name{GeoSoilQuery-package}
\alias{GeoSoilQuery-package}
\title{GeoSoilQuery-package}
\description{
Werkzeuge zur Abfrage geologischer Parameter aus dem IGME5000-Datensatz
auf Basis von Punkt oder Batch Koordinaten.
}
\details{
Das Paket enthält:
\itemize{
\item die R6-Klasse \code{GeoSoilQuery}
\item Wrapper-Funktionen für einfache Nutzung
\item optional mitgelieferte Beispieldaten (inst/extdata)
}

Die Geodaten stammen aus dem IGME5000-Projekt und werden hier ausschließlich
als Beispieldaten für Demonstrations- und Testzwecke verwendet.
}
\keyword{"_PACKAGE"}
\keyword{internal}
